refactor(nightscout): move `GlucoseThreshold` into nightscout models, few renames for clarity
