+++ src/conversions/glucose.rs
@@ -1,5 +1,6 @@
 use crate::util::deserializers::deserialize_glucose;
 use crate::util::math::round_to;
+use crate::util::nightscout::v1_models::GlucoseThreshold;
 use serde::{Deserialize, Serialize};
 use std::cmp::Ordering;
 use std::fmt;
@@ -83,8 +84,8 @@ impl Glucose {
     }
 
     /// Returns the glucose status based on the given threshold.
-    pub fn get_status(&self, threshold: &GlucoseThreshold) -> GlucoseStatus {
-        threshold.get_status(self)
+    pub fn status(&self, threshold: &GlucoseThreshold) -> GlucoseStatus {
+        GlucoseStatus::new_status(self, threshold)
     }
 }
 
@@ -155,34 +156,16 @@ pub enum GlucoseStatus {
     Inside,
 }
 
-/// Glucose thresholds (high, target top, target bottom, low).
-///
-/// Matches the `thresholds` object from Nightscout API endpoint: `/api/v1/status.json`
-#[derive(Debug, Clone, Copy, Deserialize)]
-pub struct GlucoseThreshold {
-    #[serde(rename = "bgHigh")]
-    pub bg_high: Glucose,
-
-    #[serde(rename = "bgTargetTop")]
-    pub bg_target_top: Glucose,
-
-    #[serde(rename = "bgTargetBottom")]
-    pub bg_target_bottom: Glucose,
-
-    #[serde(rename = "bgLow")]
-    pub bg_low: Glucose,
-}
-
-impl GlucoseThreshold {
+impl GlucoseStatus {
     /// Returns the glucose status based on the threshold ranges.
     ///
     /// * [`GlucoseStatus::Urgent`] if the glucose is below `bg_low` or above `bg_high`.
     /// * [`GlucoseStatus::Outside`] if the glucose is outside the target range but within high/low bounds.
     /// * [`GlucoseStatus::Inside`] if the glucose is within the target range.
-    pub fn get_status(&self, glucose: &Glucose) -> GlucoseStatus {
-        if *glucose >= self.bg_high || *glucose <= self.bg_low {
+    pub fn new_status(glucose: &Glucose, threshold: &GlucoseThreshold) -> GlucoseStatus {
+        if *glucose >= threshold.bg_high || *glucose <= threshold.bg_low {
             GlucoseStatus::Urgent
-        } else if *glucose >= self.bg_target_top || *glucose <= self.bg_target_bottom {
+        } else if *glucose >= threshold.bg_target_top || *glucose <= threshold.bg_target_bottom {
             GlucoseStatus::Outside
         } else {
             GlucoseStatus::Inside
